+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Subtype from XAMReader and XAMWriter from common abstract types.
+- Subtype from XAMRecord.
+- Unified flag queries.
 
 ## [0.3.1]